backend/requirements.txt
flask
flask-cors
gunicorn
requests
python-dotenv
yt-dlp
